requirements.txt
PyQt5>=5.15
PyQtWebEngine>=5.15
folium>=0.14
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pyproj>=3.5
shapely>=2.0
rasterio>=1.3
Pillow>=9.0
pandas>=2.0
pulp>=2.7
scikit-learn>=1.3